+++ R/show_forest.R
@@ -9,6 +9,7 @@
 #' @param vars_to_show default is `NULL`, show all variables
 #' (including controls). You can use this to choose variables
 #' to show, but remember, the models have not been changed.
+#' @param add_caption if `TRUE`, add caption to the plot.
 #'
 #' @return a `ggplot` object
 #' @export
@@ -32,21 +33,37 @@ show_forest <- function(data, covariates, controls = NULL,
                         model_names = NULL,
                         vars_to_show = NULL,
                         drop_controls = FALSE,
+                        add_caption = TRUE,
+                        headings = list(variable = "Variable", n = "N", measure = "Hazard ratio", ci = NULL, p = "p"),
                         model_dir = file.path(tempdir(), "ezcox"),
+                        verbose = TRUE,
                         ...) {
   fit_df <- ezcox(data,
     covariates = covariates,
     controls = controls,
     time = time, status = status,
     model_dir = model_dir,
-    return_models = TRUE
+    return_models = TRUE,
+    verbose = verbose
   )
   fit_models <- get_models(fit_df)
-  show_models(fit_models,
+  p <- show_models(fit_models,
     model_names = model_names,
     covariates = vars_to_show,
     merge_models = merge_models,
     drop_controls = drop_controls,
+    headings = headings,
     ...
   )
+
+  if (add_caption) {
+    if (is.null(controls)) {
+      p <- p + ggplot2::labs(caption = paste("Cox analysis for variable", paste(covariates, collapse = " & ")))
+    } else {
+      p <- p + ggplot2::labs(caption = paste("Cox analysis for variable", paste(covariates, collapse = " & "), "\n",
+                                             "with", paste(controls, collapse = " & "), "controlled"))
+    }
+  }
+
+  p
 }